[Fix] : database connection error fix
Git : #21
