+++ eb_fast_api.Dockerfile
@@ -1,9 +1,10 @@
 FROM python:3.12.3-slim
 
+WORKDIR /api/
 COPY . .
 
 RUN pip install -e .
 RUN pip install -r requirements.txt
-WORKDIR /eb_fast_api/
+WORKDIR /api/eb_fast_api/
 
-ENTRYPOINT [ "uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000", "--reload" ]
+ENTRYPOINT [ "python", "main.py" ]
